Added rolls.py, moved dice functions. Cleaned up dice.py
